Hide [!DNL x] and [!UICONTROL x] tags from Preview Fixes #12
